Added grid plane

+++ OpenGL/src/Application.h
@@ -3,7 +3,6 @@
 #include <GL/glew.h>
 #include "Window.h"
 #include "Render.h"
-#include <string>
 
 class Application
 {
@@ -19,8 +18,8 @@ private:
 	float cubeAngle = 0.0f;
 
 	int frames = 0;
-	double currentTime;
-	double deltaTime = 0.0;
-	double elapsedTime = 0.0;
-	double previousTime = 0.0;
+	float currentTime;
+	float deltaTime = 0.0;
+	float elapsedTime = 0.0;
+	float previousTime = 0.0;
 };
